+++ oshi-core/src/main/java/oshi/hardware/platform/linux/LinuxNetworkIF.java
@@ -59,26 +59,16 @@ public final class LinuxNetworkIF extends AbstractNetworkIF {
     }
 
     /**
-     * Gets non-local network interfaces on this machine
-     *
-     * @return An {@code UnmodifiableList} of {@link NetworkIF} objects representing
-     *         the interfaces
-     */
-    public static List<NetworkIF> getNetworks() {
-        return Collections.unmodifiableList(
-            getNetworkInterfaces(false).stream().map(LinuxNetworkIF::new).collect(Collectors.toList()));
-    }
-
-    /**
      * Gets network interfaces on this machine
      *
-     * @param includeLocalInterfaces include local interfaces in the result
+     * @param includeLocalInterfaces
+     *            include local interfaces in the result
      * @return An {@code UnmodifiableList} of {@link NetworkIF} objects representing
      *         the interfaces
      */
     public static List<NetworkIF> getNetworks(boolean includeLocalInterfaces) {
-        return Collections.unmodifiableList(
-                getNetworkInterfaces(includeLocalInterfaces).stream().map(LinuxNetworkIF::new).collect(Collectors.toList()));
+        return Collections.unmodifiableList(getNetworkInterfaces(includeLocalInterfaces).stream()
+                .map(LinuxNetworkIF::new).collect(Collectors.toList()));
     }
 
     @Override